requirements.txt
torch>=2.2
torchvision>=0.17
networkx>=3.2
faiss-cpu>=1.7.4
pydantic>=2.4
hydra-core>=1.3
numpy>=1.26
scipy>=1.11
pytest>=8.0
transformers>=4.36
datasets>=2.14
ollama>=0.1.0
requests>=2.31
pillow>=10.0
tqdm>=4.66
torch-geometric>=2.4.0
sentence-transformers>=2.2.0
